move Button.kt and Operator.kt file to util package and delete homeScreen package
