+++ app/src/main/java/com/sadegh/calculator/presentation/main_screen/components/Buttons.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.sadegh.calculator.homeScreen.Button
+import com.sadegh.calculator.presentation.main_screen.util.Button
 import com.sadegh.calculator.presentation.main_screen.UserEvent
 
 @Composable
